pyproject.toml
[project]
name = "industrial-robot-predictive-mtce"
version = "0.1.0"
description = "End-to-end predictive maintenance system for industrial six-axis robots"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Hitesh", email = "hjha03144@gmail.com"}
]
keywords = ["predictive-maintenance", "machine-learning", "xgboost", "streamlit", "robotics"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "streamlit>=1.30.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=12.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "joblib>=1.3",
    "plotly>=5.15",
    "psutil>=5.9",
    "tqdm",
    "pyyaml",
    "click",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
robotguard = "app.main:main"               # if you want CLI entrypoint later
train-models = "src.modeling.train:main"
generate-data = "src.etl.synthetic_data_generator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "app"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "PL", "UP", "B", "C4"]
ignore = ["E501"]  # line too long – handled by black

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]